# contracts/openapi/universal-runtime-phase1.yaml
openapi: 3.1.0
info:
  title: UniversalRuntime Phase 1 API
  version: 1.0.0-alpha.1
  description: Native management API plus selected LangGraph-compatible resource paths.
servers:
  - url: http://localhost:8080
paths:
  /ok:
    get:
      operationId: health
      responses:
        '200':
          description: Healthy
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
  /info:
    get:
      operationId: info
      responses:
        '200':
          description: Runtime information
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RuntimeInfo' }
  /api/v1/applications/{application_id}/config:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      operationId: getApplicationConfig
      responses:
        '200':
          description: Active config revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NativeResponse' }
    put:
      operationId: createApplicationConfigRevision
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201':
          description: Immutable revision created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NativeResponse' }
  /api/v1/applications/{application_id}/config/validate:
    post:
      operationId: validateApplicationConfig
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Validation result
  /api/v1/applications/{application_id}/config/{revision}/activate:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - name: revision
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    post:
      operationId: activateApplicationConfigRevision
      responses:
        '200':
          description: Activated immutable revision
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NativeResponse' }
  /assistants:
    post:
      operationId: createAssistant
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssistantCreate' }
      responses:
        '200':
          description: LangGraph-compatible assistant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assistant' }
  /assistants/search:
    post:
      operationId: searchAssistants
      responses:
        '200':
          description: Search result
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Assistant' }
  /assistants/{assistant_id}:
    get:
      operationId: getAssistant
      parameters:
        - $ref: '#/components/parameters/AssistantId'
      responses:
        '200':
          description: Assistant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assistant' }
  /assistants/{assistant_id}/schemas:
    get:
      operationId: getAssistantSchemas
      parameters:
        - $ref: '#/components/parameters/AssistantId'
      responses:
        '200':
          description: Graph schemas
  /threads:
    post:
      operationId: createThread
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ThreadCreate' }
      responses:
        '200':
          description: Thread
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Thread' }
  /threads/search:
    post:
      operationId: searchThreads
      responses:
        '200':
          description: Threads
  /threads/{thread_id}:
    get:
      operationId: getThread
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '200':
          description: Thread
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Thread' }
  /threads/{thread_id}/state:
    get:
      operationId: getThreadState
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '200':
          description: Current state
    post:
      operationId: updateThreadState
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '200':
          description: Updated state
  /threads/{thread_id}/history:
    get:
      operationId: getThreadHistory
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      responses:
        '200':
          description: Checkpoint state history
  /threads/{thread_id}/runs:
    post:
      operationId: createRun
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
  /threads/{thread_id}/runs/stream:
    post:
      operationId: streamRun
      parameters:
        - $ref: '#/components/parameters/ThreadId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunCreate' }
      responses:
        '200':
          description: LangGraph-compatible SSE stream
          content:
            text/event-stream:
              schema: { type: string }
  /runs/stream:
    post:
      operationId: streamStatelessRun
      responses:
        '200':
          description: Stateless SSE stream
          content:
            text/event-stream:
              schema: { type: string }
  /runs/wait:
    post:
      operationId: waitStatelessRun
      responses:
        '200':
          description: Stateless result
  /runs/{run_id}:
    get:
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
  /runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '204': { description: Cancel requested }
  /.well-known/agent-card.json:
    get:
      operationId: getAgentCard
      responses:
        '200':
          description: A2A Agent Card generated from application/assistant metadata
          content:
            application/json:
              schema: { type: object }
components:
  parameters:
    ApplicationId:
      name: application_id
      in: path
      required: true
      schema: { type: string }
    AssistantId:
      name: assistant_id
      in: path
      required: true
      schema: { type: string }
    ThreadId:
      name: thread_id
      in: path
      required: true
      schema: { type: string }
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string }
  schemas:
    RuntimeInfo:
      type: object
      required: [version, adapters]
      properties:
        version: { type: string }
        adapters: { type: array, items: { type: object } }
    NativeResponse:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          type: object
          properties:
            request_id: { type: string }
            timestamp: { type: string, format: date-time }
    AssistantCreate:
      type: object
      properties:
        graph_id: { type: string }
        assistant_id: { type: string }
        name: { type: string }
        config: { type: object }
        context: { type: object }
        metadata: { type: object }
    Assistant:
      type: object
      required: [assistant_id, graph_id, version]
      properties:
        assistant_id: { type: string }
        graph_id: { type: string }
        version: { type: integer }
        name: { type: string }
        config: { type: object }
        context: { type: object }
        metadata: { type: object }
    ThreadCreate:
      type: object
      properties:
        thread_id: { type: string }
        metadata: { type: object }
    Thread:
      type: object
      required: [thread_id, status]
      properties:
        thread_id: { type: string }
        status: { enum: [idle, busy, interrupted, error] }
        metadata: { type: object }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    RunCreate:
      type: object
      required: [assistant_id]
      properties:
        assistant_id: { type: string }
        input: {}
        command: {}
        config: { type: object }
        context: { type: object }
        metadata: { type: object }
        stream_mode:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
        stream_subgraphs: { type: boolean }
        priority: { enum: [interactive, normal, batch] }
        multitask_strategy: { enum: [reject, enqueue, interrupt, rollback] }
    Run:
      type: object
      required: [run_id, assistant_id, status]
      properties:
        run_id: { type: string }
        thread_id: { type: [string, 'null'] }
        assistant_id: { type: string }
        status: { enum: [pending, running, interrupted, success, error, timeout, cancelled] }
        metadata: { type: object }